requirements.txt
# Add your dependencies here
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
requests
psutil
python-multipart
torchvision
Pillow
transformers
torch
sentencepiece
protobuf==3.20.*
torchaudio





